Simple Auth added

+++ routes/products_router.py
@@ -1,10 +1,13 @@
-from fastapi import APIRouter, HTTPException, Response, status
+from fastapi import APIRouter, HTTPException, Response, status, Depends
+from sqlalchemy.sql.sqltypes import String
 from schemas.product_schema import Product
 from uuid import uuid4
 from config.database import Conn
 from models.product_model import Products
 from starlette.status import HTTP_204_NO_CONTENT
+from decouple import config
 
+PASSWORD_FOR_WRITE = config('PASS_FORW')
 products = APIRouter()
 
 
@@ -18,8 +21,10 @@ def get_product(id: str):
     return Conn.execute(Products.select().where(Products.c.id == id)).first()
 
 
-@products.post('/products')
-def create_product(new_product: Product):
+@products.post('/products/{password}',)
+def create_product(new_product: Product, password:str):
+    if password != PASSWORD_FOR_WRITE:
+        raise HTTPException(status_code=401,detail="Unauthorized, wrong password.")
     n_product = {
         "name": new_product.name,
         "description": new_product.description,
@@ -29,14 +34,18 @@ def create_product(new_product: Product):
     return Conn.execute(Products.select().where(Products.c.id == r.lastrowid)).first()
 
 
-@products.delete('/products/{id}', status_code=status.HTTP_204_NO_CONTENT)
-def delete_product(id: str):
+@products.delete('/products/d/{id}{pass}', status_code=status.HTTP_204_NO_CONTENT)
+def delete_product(id: str, password:str):
+    if password != PASSWORD_FOR_WRITE:
+        raise HTTPException(status_code=401,detail="Unauthorized, wrong password.")
     Conn.execute(Products.delete().where(Products.c.id == id))
     return Response(status_code=HTTP_204_NO_CONTENT)
 
 
-@products.put('/products/{id}')
-def update_product(id: str, upd_product: Product):
+@products.put('/products/u/{id}{pass}')
+def update_product(id: str, upd_product: Product, password:str):
+    if password != PASSWORD_FOR_WRITE:
+        raise HTTPException(status_code=401,detail="Unauthorized, wrong password.")
     Conn.execute(Products.update().values(
         name=upd_product.name,
         description=upd_product.description,
